move deployment test/scripts to its own directory
