+++ migration/config/config.js
@@ -8,7 +8,7 @@ const filepath = path.join(__dirname, '../../.env');
 dotenv.config({ path: filepath });
 
 const db = {
-    database: process.env.RECREG_DB_NAME || 'recreg',
+    database: process.env.RECREG_DB_NAME || 'recregtest',
     username: process.env.RECREG_DB_USER,
     password: process.env.RECREG_DB_PASS,
     host: process.env.RECREG_DB_HOST || 'localhost',
